Change upper and lower row to ints to avoid TravisCI testing errors.

+++ tests/test_bootstrap_calcs.py
@@ -69,11 +69,11 @@ class ComputationalTests(unittest.TestCase):
         # We should have the value in BR[lower_row, 0] = 3 so that there are 2
         # elements in bootstrap_replicates (BR) that are less than this. I.e.
         # we want lower_row = 2. Note 2.56 rounded down is 2.
-        lower_row = np.floor(alpha / 2.0)
+        lower_row = int(np.floor(alpha / 2.0))
         # 100 - 2.56 is 97.44. Rounded up, this is 98.
         # We want the row such that the value in the first column of that row
         # is 98, i.e. we want the row at index 97.
-        upper_row = np.floor(100 - (alpha / 2.0))
+        upper_row = int(np.floor(100 - (alpha / 2.0)))
         # Create the expected results
         expected_results =\
             bc.combine_conf_endpoints(self.bootstrap_replicates[lower_row],
